UYVY file reader and YUV to RGB converter
These are both very rough and ready first ideas about how all this will
work. They still need a lot doing to them.

The YUV to RGB conversion is not done very cleverly, or efficiently, but
servers as an illustration of how a typical "transformer" component will
work. A lot of what's there at present will be abstracted into a base
class.

One vital missing component is a common "frame" type. It needs to
include common conversion methods and be flexible enough, but we can't
expect all components to cope with all frame types. A YUV to RGB
converter can reasonably insist on a YUV input.