Replace 'bind' with 'connect'
The 'connect' method is executed immediately, rather than being an actor
method that is executed when the component starts. It can also be
connected to any callable, rather than only to another actor input.
Lastly, it calls an 'on_connect' method for derived classes to react to
connections being made.

Signed-off-by: Jim Easterbrook <jim@jim-easterbrook.me.uk>